Update CROM & Maintain Compatibility for Production – DEV-6985
Updated CROM, pulling in latest updates from the Linked.art CROM
repository, and reinstated several previously removed model attributes
including the `Relationship` entity and its associated properties, the
`current_keeper` property and its inverse, as well as the `Geometry`
and `CoordinateSystem` entities.

These reinstated attributes are all required for ongoing compatibility
with data modeling that is in production use and will take some time to
migrate away from so as not to break several production or
near-production applications that are consuming the modeled data in its
existing form.