Merge Latest Upstream Commits – DEV-11971
Merged latest upstream `linked-art/crom` repository commits, bringing
the `thegetty/crom` up-to-date with upstream, while maintaining
backwards compatibility and support for the following upstream removed
code artifacts:

 * the `current_keeper` and `current_keeper_of` properties
 * the `Geometry` and `CoordinateSystem` entities

Support for the deprecated and no-longer used `Relationship` entity and
its `relates_to`, `related_to_by`, `relates_from`, and
`related_from_by` properties have been removed to simplify ongoing
forwards compatibility with upstream.

When we are able to migrate away from the use of the `current_keeper`
property in the Museum LOD, and the use of the `Geometry` and
`CoordinateSystem` entities in the Vocabulary LOD, we should be able to
switch back to the use of the upstream repository and upstream
Linked.art context, which are becoming more stable as they mature.